Move event under eth module
