+++ src/eth/mod.rs
@@ -1,3 +1,6 @@
 pub mod contracts;
 pub mod transaction;
 pub mod utils;
+pub mod event;
+
+pub use event::Event;
